// app/src/@shared/enums/ScreenNames.enum.ts
export enum Screens {
	// Bottom tabs Screen
	HOME_TABS = "HOME_TABS",
	CART_TABS = "CART_TABS",
	WISHLIST_TABS = "WISHLIST_TABS",
	PROFILE_TABS = "PROFILE_TABS",

	// Drawer Screens
	HOME_DRAWER = "HOME_DRAWER",
	SPLASH_DRAWER = "SPLASH_DRAWER",
	SIGN_UP_DRAWER = "SIGN_UP_DRAWER",
	WISHLIST_DRAWER = "WISHLIST_DRAWER",
	CART_DRAWER = "CART_DRAWER",
	PRODUCTS_DRAWER = "PRODUCTS_DRAWER",
	PRODUCT_DETAILS_DRAWER = "PRODUCT_DETAILS_DRAWER",
	PROFILE_DRAWER = "PROFILE_DRAWER",
	UNDER_CONSTRICTION_DRAWER = "UNDER_CONSTRICTION_DRAWER",

	ECOMMARCE_HOME = "ECOMMARCE_HOME",
	PAYMENT = "PAYMENT",
	ADD_NEW_CARD = "ADD_NEW_CARD_SCREEN",
	PROFILE_SCREEN = "PROFILE_SCREEN",
	PROFILE_UPDATE_SCREEN = "PROFILE_UPDATE_SCREEN",
	CHAT_SCREEN = "CHAT_SCREEN",
	CONVERSATION_LIST_SCREEN = "CONVERSATION_LIST_SCREEN",
	FEED_SCREEN = "FEED_SCREEN",
	SWIPE_SCREEN = "SWIPE_SCREEN",
	POST_DETAILS_SCREEN = "POST_DETAILS_SCREEN",
	PET_LISTING_SCREEN = "PET_LISTING_SCREEN",

	SIGN_IN = "SIGN_IN",
}